Pre 'symbol notation change' commit
While processing 'EQUATION' kinds of variables I've encountered the following problem:
All my Symbol variables strings are written in latex notation. Content of 'EQUATION' variable field is also corresponds to this notation. But it seems that there is no good available parser for translating from latex style string to sympy equation.
Solution:
- change notation of all my symbols to sympy.
Possible consequences:
- this should not affect rendering of sympy equations into latex (sympy->latex translation is ok but not reversable as noticed above).
- some special simbols like \alpha may arise a problem during sympy_like_str -> sympy parsing with parse_expr.

This commit is intended to save huge work and should be used as a target of the 'git reverse commit' action in case it will not be possible to solve the problem with this solution.